Created setAPN function in cInent, changed FirstTimeInetConfig to InetConfig, and added setAPN call to InetConfig.
